+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [PR 156](https://github.com/athelas-astro/athelas/pull/156) Clean up total energy and momentum history.
 
 ### Fixed (not changing behavior/API/variables/...)
+- [PR 160](https://github.com/athelas-astro/athelas/pull/160) Progress on weno limiter.
 - [PR 155](https://github.com/athelas-astro/athelas/pull/155) Add radiation energy to thermal engine calibration.
 - [PR 154](https://github.com/athelas-astro/athelas/pull/154) Fix to thermal engine calibration.